# swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Forecasting Platform API",
    "description": "A simplified Polymarket-like forecasting platform with idempotency, state machine control, and ledger reconciliation.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
    "/api/users/{id}/deposit": {
      "post": {
        "summary": "Deposit funds",
        "description": "Add funds to a user's balance with idempotency support",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "User ID (e.g., 1, 2, 3)"
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Required string for idempotency"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount"],
                "properties": {
                  "amount": {
                    "type": "number",
                    "minimum": 0,
                    "example": 100,
                    "description": "Amount to deposit"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deposit successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "username": {
                      "type": "string"
                    },
                    "balance": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid amount or missing Idempotency-Key"
          },
          "409": {
            "description": "Idempotency key already used with different amount"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/bets": {
      "post": {
        "summary": "Place a bet",
        "description": "Place a bet on a game with idempotency support",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Required string for idempotency"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["userId", "gameId", "amount"],
                "properties": {
                  "userId": {
                    "type": "integer",
                    "example": 68,
                    "description": "User ID (e.g., 68, 69, 70)"
                  },
                  "gameId": {
                    "type": "string",
                    "description": "Game identifier"
                  },
                  "amount": {
                    "type": "number",
                    "minimum": 0,
                    "description": "Bet amount"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bet placed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "userId": {
                      "type": "integer"
                    },
                    "gameId": {
                      "type": "string"
                    },
                    "amount": {
                      "type": "number"
                    },
                    "status": {
                      "type": "string",
                      "enum": ["PLACED"]
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or insufficient balance"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/bets/{id}/settle": {
      "post": {
        "summary": "Settle a bet",
        "description": "Settle a bet as WIN or LOSE",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Required string for idempotency"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Bet ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["result"],
                "properties": {
                  "result": {
                    "type": "string",
                    "enum": ["WIN", "LOSE"],
                    "description": "Settlement result"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bet settled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "userId": {
                      "type": "integer"
                    },
                    "gameId": {
                      "type": "string"
                    },
                    "amount": {
                      "type": "number"
                    },
                    "status": {
                      "type": "string",
                      "enum": ["SETTLED"]
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid result or bet not in PLACED status"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/bets/{id}/cancel": {
      "post": {
        "summary": "Cancel a bet",
        "description": "Cancel a bet and refund the amount",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Required string for idempotency"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Bet ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Bet cancelled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "userId": {
                      "type": "integer"
                    },
                    "gameId": {
                      "type": "string"
                    },
                    "amount": {
                      "type": "number"
                    },
                    "status": {
                      "type": "string",
                      "enum": ["CANCELLED"]
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bet not in PLACED status"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/admin/reconcile": {
      "get": {
        "summary": "Admin reconciliation",
        "description": "Verify account balance and detect anomalies",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "User ID to reconcile"
          }
        ],
        "responses": {
          "200": {
            "description": "Reconciliation successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "integer"
                    },
                    "currentBalance": {
                      "type": "string"
                    },
                    "calculatedBalance": {
                      "type": "string"
                    },
                    "isConsistent": {
                      "type": "boolean"
                    },
                    "betStats": {
                      "type": "object",
                      "properties": {
                        "placed": {
                          "type": "integer"
                        },
                        "settled": {
                          "type": "integer"
                        },
                        "cancelled": {
                          "type": "integer"
                        }
                      }
                    },
                    "anomalies": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  }
}